ath79: update GL-AR300M dts to detect nand flash
GL-AR300M has two types of flash both on the SPI bus, when cs-num is set
to 1 it will only detect the first flash chip. As mentioned by Robert[1]
we cannot just set cs-num to 2 without defining where the pins are. He
suggested trying hardware GPIO's but I was unable to find one that
worked, as described in ath79's ar300m initial pull request[2] we can use
native cs-gpios. This didn't work yet because we don't have the
required chip id's in the kernel to actually detect the flash chip used
in GL-AR300M.

The kernel patch and UBIFS related config changes will be provided as
separate commits.

  [1] https://bugs.openwrt.org/index.php?do=details&task_id=1782
  [2] https://github.com/openwrt/openwrt/pull/983

Signed-off-by: Menno <mennozon@gmail.com>